requirements.txt
asgiref==3.2.10
boto3==1.14.30
botocore==1.17.30
certifi==2020.6.20
chardet==3.0.4
Django==3.0.8
django-templated-mail==1.1.1
djangorestframework==3.11.0
djoser==2.0.3
docutils==0.15.2
idna==2.10
jmespath==0.10.0
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
twilio==6.44.1
urllib3==1.25.10
